Add full scale sort by rank tests
