+++ core/src/test/scala/it/parttimeteam/core/TestCards.scala
@@ -13,6 +13,7 @@ object TestCards {
   val ACE_CLUBS_BLUE: Card = Card(Rank.Ace(), Suit.Clubs(), Color.Blue())
   val ACE_CLUBS_RED: Card = Card(Rank.Ace(), Suit.Clubs(), Color.Red())
   val OVERFLOW_ACE_CLUBS: Card = Card(Rank.OverflowAce(), Suit.Clubs(), Color.Blue())
+  val OVERFLOW_ACE_SPADES: Card = Card(Rank.OverflowAce(), Suit.Spades(), Color.Blue())
   val ACE_SPADES: Card = Card(Rank.Ace(), Suit.Spades(), Color.Blue())
   val ACE_DIAMONDS: Card = Card(Rank.Ace(), Suit.Diamonds(), Color.Red())
   val ACE_HEARTS: Card = Card(Rank.Ace(), Suit.Hearts(), Color.Red())
@@ -53,23 +54,27 @@ object TestCards {
    */
   val SIX_HEARTS: Card = Card(Rank.Six(), Suit.Hearts(), Color.Red())
   val SIX_CLUBS: Card = Card(Rank.Six(), Suit.Clubs(), Color.Red())
+  val SIX_SPADES: Card = Card(Rank.Six(), Suit.Spades(), Color.Red())
 
   /**
    * Seven cards.
    */
   val SEVEN_HEARTS: Card = Card(Rank.Seven(), Suit.Hearts(), Color.Red())
   val SEVEN_CLUBS: Card = Card(Rank.Seven(), Suit.Clubs(), Color.Red())
+  val SEVEN_SPADES: Card = Card(Rank.Seven(), Suit.Spades(), Color.Red())
 
   /**
    * Eight cards.
    */
   val EIGHT_CLUBS: Card = Card(Rank.Eight(), Suit.Clubs(), Color.Red())
+  val EIGHT_SPADES: Card = Card(Rank.Eight(), Suit.Spades(), Color.Red())
 
   /**
    * Nine cards.
    */
   val NINE_HEARTS: Card = Card(Rank.Nine(), Suit.Hearts(), Color.Red())
   val NINE_CLUBS: Card = Card(Rank.Nine(), Suit.Clubs(), Color.Red())
+  val NINE_SPADES: Card = Card(Rank.Nine(), Suit.Spades(), Color.Red())
 
   /**
    * Ten cards.
@@ -84,6 +89,7 @@ object TestCards {
    */
   val JACK_HEARTS: Card = Card(Rank.Jack(), Suit.Hearts(), Color.Blue())
   val JACK_CLUBS: Card = Card(Rank.Jack(), Suit.Clubs(), Color.Blue())
+  val JACK_SPADES: Card = Card(Rank.Jack(), Suit.Spades(), Color.Blue())
 
   /**
    * Queen cards.
